+++ src/lsp-server/schema.ts
@@ -34,7 +34,7 @@ function setSchema(schemaConfig: { versionName: string; customPath: string }) {
         break;
       case "Custom":
         console.log(
-          "Selected custom schema, but no path provided.  Setting to default.",
+          "Selected custom schema, but no path provided.  Setting to default."
         );
         schemaPath = path.join(schemaDir, "pretext.rng");
         break;
@@ -70,7 +70,7 @@ export class Schema {
           if (nodeName) {
             aliasMap[nodeName] = deepmerge(
               aliasMap[nodeName] || {},
-              getChildren(node),
+              getChildren(node)
             );
           }
         } else if (node.name === "element") {
@@ -78,7 +78,7 @@ export class Schema {
           if (nodeName) {
             tmpElementChildren[nodeName] = deepmerge(
               tmpElementChildren[nodeName] || {},
-              getChildren(node),
+              getChildren(node)
             );
           }
         }